Move Data Update Coordinator to the coordinator module
