+++ homeassistant/components/nam/__init__.py
@@ -2,17 +2,13 @@
 
 from __future__ import annotations
 
-import asyncio
 import logging
-from typing import cast
 
 from aiohttp.client_exceptions import ClientConnectorError, ClientError
 from nettigo_air_monitor import (
     ApiError,
     AuthFailedError,
     ConnectionOptions,
-    InvalidSensorDataError,
-    NAMSensors,
     NettigoAirMonitor,
 )
 
@@ -21,25 +17,20 @@ from homeassistant.config_entries import ConfigEntry
 from homeassistant.const import CONF_HOST, CONF_PASSWORD, CONF_USERNAME, Platform
 from homeassistant.core import HomeAssistant
 from homeassistant.exceptions import ConfigEntryAuthFailed, ConfigEntryNotReady
-from homeassistant.helpers import device_registry as dr, entity_registry as er
+from homeassistant.helpers import entity_registry as er
 from homeassistant.helpers.aiohttp_client import async_get_clientsession
-from homeassistant.helpers.device_registry import DeviceInfo
-from homeassistant.helpers.update_coordinator import DataUpdateCoordinator, UpdateFailed
 
-from .const import (
-    ATTR_SDS011,
-    ATTR_SPS30,
-    DEFAULT_UPDATE_INTERVAL,
-    DOMAIN,
-    MANUFACTURER,
-)
+from .const import ATTR_SDS011, ATTR_SPS30, DOMAIN
+from .coordinator import NAMDataUpdateCoordinator
 
 _LOGGER = logging.getLogger(__name__)
 
 PLATFORMS = [Platform.BUTTON, Platform.SENSOR]
 
+NAMConfigEntry = ConfigEntry[NAMDataUpdateCoordinator]
 
-async def async_setup_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
+
+async def async_setup_entry(hass: HomeAssistant, entry: NAMConfigEntry) -> bool:
     """Set up Nettigo as config entry."""
     host: str = entry.data[CONF_HOST]
     username: str | None = entry.data.get(CONF_USERNAME)
@@ -63,8 +54,7 @@ async def async_setup_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
     coordinator = NAMDataUpdateCoordinator(hass, nam, entry.unique_id)
     await coordinator.async_config_entry_first_refresh()
 
-    hass.data.setdefault(DOMAIN, {})
-    hass.data[DOMAIN][entry.entry_id] = coordinator
+    entry.runtime_data = coordinator
 
     await hass.config_entries.async_forward_entry_setups(entry, PLATFORMS)
 
@@ -81,57 +71,6 @@ async def async_setup_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
     return True
 
 
-async def async_unload_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
+async def async_unload_entry(hass: HomeAssistant, entry: NAMConfigEntry) -> bool:
     """Unload a config entry."""
-    unload_ok = await hass.config_entries.async_unload_platforms(entry, PLATFORMS)
-
-    if unload_ok:
-        hass.data[DOMAIN].pop(entry.entry_id)
-
-    return unload_ok
-
-
-class NAMDataUpdateCoordinator(DataUpdateCoordinator[NAMSensors]):  # pylint: disable=hass-enforce-coordinator-module
-    """Class to manage fetching Nettigo Air Monitor data."""
-
-    def __init__(
-        self,
-        hass: HomeAssistant,
-        nam: NettigoAirMonitor,
-        unique_id: str | None,
-    ) -> None:
-        """Initialize."""
-        self._unique_id = unique_id
-        self.nam = nam
-
-        super().__init__(
-            hass, _LOGGER, name=DOMAIN, update_interval=DEFAULT_UPDATE_INTERVAL
-        )
-
-    async def _async_update_data(self) -> NAMSensors:
-        """Update data via library."""
-        try:
-            async with asyncio.timeout(10):
-                data = await self.nam.async_update()
-        # We do not need to catch AuthFailed exception here because sensor data is
-        # always available without authorization.
-        except (ApiError, ClientConnectorError, InvalidSensorDataError) as error:
-            raise UpdateFailed(error) from error
-
-        return data
-
-    @property
-    def unique_id(self) -> str | None:
-        """Return a unique_id."""
-        return self._unique_id
-
-    @property
-    def device_info(self) -> DeviceInfo:
-        """Return the device info."""
-        return DeviceInfo(
-            connections={(dr.CONNECTION_NETWORK_MAC, cast(str, self._unique_id))},
-            name="Nettigo Air Monitor",
-            sw_version=self.nam.software_version,
-            manufacturer=MANUFACTURER,
-            configuration_url=f"http://{self.nam.host}/",
-        )
+    return await hass.config_entries.async_unload_platforms(entry, PLATFORMS)
